Support kotlin.time.Instant in FormattedInstantSerializer
The DateTimeFormat-based FormattedInstantSerializer would be
introduced in this release for `kotlinx.datetime.Instant` if that
wasn't going to be removed.
Now, instead, we support the new, `kotlin.time.Instant` in this
serializer.
