+++ integration-testing/serialization/common/test/DeprecatedInstantSerializationTest.kt
@@ -7,7 +7,6 @@
 package kotlinx.datetime.serialization.test
 
 import kotlinx.datetime.*
-import kotlinx.datetime.format.DateTimeComponents
 import kotlinx.datetime.serializers.*
 import kotlinx.serialization.*
 import kotlinx.serialization.json.*
@@ -87,27 +86,5 @@ class DeprecatedInstantSerializationTest {
         // should be the same as the ISO 8601
         assertKSerializerName<Instant>("kotlinx.datetime.Instant", Json.serializersModule.serializer())
         defaultSerialization(Json.serializersModule.serializer())
-    }
-
-    object Rfc1123InstantSerializer : FormattedInstantSerializer("RFC_1123", DateTimeComponents.Formats.RFC_1123)
-
-    @Test
-    fun testCustomSerializer() {
-        assertKSerializerName("kotlinx.datetime.Instant/serializer/RFC_1123", Rfc1123InstantSerializer)
-        for ((instant, json) in listOf(
-            Pair(Instant.fromEpochSeconds(1607505416),
-                "\"Wed, 9 Dec 2020 09:16:56 GMT\""),
-            Pair(Instant.fromEpochSeconds(-1607505416),
-                "\"Thu, 23 Jan 1919 14:43:04 GMT\""),
-            Pair(Instant.fromEpochSeconds(987654321),
-                "\"Thu, 19 Apr 2001 04:25:21 GMT\""),
-        )) {
-            assertEquals(json, Json.encodeToString(Rfc1123InstantSerializer, instant))
-            assertEquals(instant, Json.decodeFromString(Rfc1123InstantSerializer, json))
-        }
-        assertEquals("\"Thu, 19 Apr 2001 04:25:21 GMT\"",
-            Json.encodeToString(Rfc1123InstantSerializer, Instant.fromEpochSeconds(987654321, 123456789)))
-        assertEquals(Instant.fromEpochSeconds(987654321),
-            Json.decodeFromString(Rfc1123InstantSerializer, "\"Thu, 19 Apr 2001 08:25:21 +0400\""))
     }
 }
